move common files to /shared
