+++ packages/components/src/DateTimePickers/Date/Manager/Manager.component.js
@@ -78,7 +78,6 @@ function ContextualManager(props) {
 ContextualManager.propTypes = {
 	children: PropTypes.node,
 	dateFormat: PropTypes.string,
-	// eslint-disable-next-line react/no-unused-prop-types
 	onChange: PropTypes.func,
 	value: PropTypes.oneOfType([PropTypes.instanceOf(Date), PropTypes.number, PropTypes.string]),
 	useUTC: PropTypes.bool,
