+++ packages/components/src/DateTimePickers/InputDateTimePicker/InputDateTimePicker.scss
@@ -1,5 +1,5 @@
 .date-time-picker {
 	display: flex;
-    flex-direction: row;
-    align-items: center;
+	flex-direction: row;
+	align-items: center;
 }
