+++ packages/components/src/DateTimePickers/shared/error-messages.js
@@ -1,4 +1,4 @@
-/* eslint-disable import/no-named-as-default-member */
+// eslint-disable-next line import/no-named-as-default-member
 import i18next from 'i18next';
 
 export default function getErrorMessage(key) {
@@ -28,6 +28,10 @@ export default function getErrorMessage(key) {
 		case 'INVALID_DAY_OF_MONTH':
 			return i18next.t('INVALID_DAY_OF_MONTH', {
 				defaultValue: "Day value doesn't match an existing day in the month",
+			});
+		case 'INVALID_TIME_EMPTY':
+			return i18next.t('INVALID_TIME_EMPTY', {
+				defaultValue: 'Time is required',
 			});
 		case 'TIME_FORMAT_INVALID':
 			return i18next.t('TIME_FORMAT_INVALID', { defaultValue: 'Time is invalid' });
